pyproject.toml
[project]
name = "deepsight"
version = "0.1.0"
description = ""
authors = [
    { name = "Francesco Gentile", email = "francesco.gentile@studenti.unitn.it" },
    { name = "Diego Calanzone", email = "diego.calanzone@studenti.unitn.it" },
]
dependencies = [
    "torch>=2.0.1",
    "torchvision>=0.15.2",
    "jaxtyping>=0.2.20",
    "torchmetrics>=0.11.4",
    "pyserde>=0.11.0",
    "numpy>=1.25.0",
    "rustworkx>=0.13.0",
    "transformers>=4.30.2",
    "scikit-learn>=1.3.0",
    "ruamel-yaml>=0.17.32",
    "wandb>=0.15.4",
    "albumentations>=1.3.1",
    "openai>=0.27.8",
    "ultralytics>=8.0.131",
    "torch-scatter>=2.1.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
lint = ["black>=23.3.0", "ruff>=0.0.272", "mypy>=1.3.0", "pre-commit>=3.3.2"]
test = ["ipykernel>=6.23.2"]
docs = ["sphinx>=7.0.1"]
types = ["types-Pillow>=9.5.0.5", "types-tqdm>=4.65.0.1"]

[[tool.pdm.source]]
type = "find_links"
url = "https://data.pyg.org/whl/torch-2.0.1+cu117.html"
name = "torch-scatter"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
select = ["F", "E", "I", "W"]
fix = true
ignore = ["F722"]
